fight_detection_using_crnn/requirement.txt.txt
cudatoolkit               10.0.130   (for tensorflow-gpu only)              
cudnn                     7.6.0  
imutils                   0.5.3 
jsonschema                3.1.1  
keras                     2.3.1
matplotlib                3.1.2
numpy                     1.16.5
opencv                    4.2.0
pillow                    6.2.1
pip                       19.3.1
python                    3.7.0 (my version will work fine in 
						> 3.6)
python-dateutil           2.8.0 
requests                  2.23.0  
tensorflow-gpu            2.0.0 (if you want to train the 							model)
other wise  tensorflow                2.0.0 
tqdm                      4.44.1 
argparse
